// src/auth/entities/usuario.entity.d.ts
export declare class Usuario {
    id_usuario: number;
    password: string;
    username: string;
    dni: number;
    apellido: string;
    nombre: string;
    roles: string[];
    id_institucion: number;
    id_equipot: number;
    fecha_alta: Date;
    ultima_actualizacion: Date;
    fecha_baja: Date;
}